Add files via upload
How can I ensure that my code, program, or software is functional and secure?

Developers should work in an environment that recommends specific patterns and promotes security best practices through environment notifications to guarantee that code, programs, or software is functional and secure. Additionally, they should base their choice of programming language and framework on the demands placed on the project and their organization. Developers must know the advantages and disadvantages of the selected language to appreciate the difficulties fully.
Finally, developers should automate their workflow to make adopting security best practices simpler.
Instead of performing security scans solely at the end of the process, the intention is to encourage developers to address issues as they arise throughout the ordinary development course. 

How do I interpret user needs and incorporate them into a program?

User needs are specifications that enhance a user's experience with a product, service, or environment. We must engage in a process where consumers comprehend their issues, procedures, objectives, and preferences to include user needs in a program. Examples include functions, which are the tasks that users must complete. Functionality directs the developers' choices for what to create and use. Next is accessibility; user requirements should be gathered from users with various traits and skills. The final category is featured, related to functions because it describes how procedures are put into practice.

How do I approach designing software?
When it comes to software design, we do have two methods. One is the "bottom-up" method, in which the design process begins with the minor possible components and subsystems. These components serve as the building blocks for the following immediately higher-level components and subsystems. This method can be combined with the top-down methodology to mask the implementation's low-level specifics. The top-down technique is the next, in which each system is broken down into some subsystems and components. Each subsystem is then broken down into a group of related subsystems and components. This division method makes it easier to create a hierarchy within the system. The primary benefit of a top-down approach is that it places a significant emphasis on requirements, which makes a design responsive to those requirements.